Добавил тесты
